more methods in the Network class
