+++ src/pyhgf/model/hgf.py
@@ -1,17 +1,13 @@
 # Author: Nicolas Legrand <nicolas.legrand@cas.au.dk>
 
-from typing import Callable, Dict, List, Optional, Tuple, Union
+from typing import Callable, Dict, Optional, Tuple, Union
 
 import jax.numpy as jnp
 import numpy as np
-import pandas as pd
 from jax.typing import ArrayLike
 
 from pyhgf.model import Network
-from pyhgf.networks import fill_categorical_state_node, to_pandas
-from pyhgf.plots import plot_correlations, plot_network, plot_nodes, plot_trajectories
 from pyhgf.response import first_level_binary_surprise, first_level_gaussian_surprise
-from pyhgf.typing import AdjacencyLists, Inputs, input_types
 
 
 class HGF(Network):
@@ -226,22 +222,6 @@ class HGF(Network):
             # initialize the model so it is ready to receive new observations
             self.create_belief_propagation_fn()
 
-    def plot_nodes(self, node_idxs: Union[int, List[int]], **kwargs):
-        """Plot the node(s) beliefs trajectories."""
-        return plot_nodes(hgf=self, node_idxs=node_idxs, **kwargs)
-
-    def plot_trajectories(self, **kwargs):
-        """Plot the parameters trajectories."""
-        return plot_trajectories(hgf=self, **kwargs)
-
-    def plot_correlations(self):
-        """Plot the heatmap of cross-trajectories correlation."""
-        return plot_correlations(hgf=self)
-
-    def plot_network(self):
-        """Visualization of node network using GraphViz."""
-        return plot_network(hgf=self)
-
     def surprise(
         self,
         response_function: Optional[Callable] = None,
@@ -289,369 +269,3 @@ class HGF(Network):
             response_function_inputs=response_function_inputs,
             response_function_parameters=response_function_parameters,
         )
-
-    def to_pandas(self) -> pd.DataFrame:
-        """Export the nodes trajectories and surprise as a Pandas data frame.
-
-        Returns
-        -------
-        structure_df :
-            Pandas data frame with the time series of sufficient statistics and
-            the surprise of each node in the structure.
-
-        """
-        return to_pandas(self)
-
-    def add_nodes(
-        self,
-        kind: str = "continuous-state",
-        n_nodes: int = 1,
-        node_parameters: Dict = {},
-        value_children: Optional[Union[List, Tuple, int]] = None,
-        value_parents: Optional[Union[List, Tuple, int]] = None,
-        volatility_children: Optional[Union[List, Tuple, int]] = None,
-        volatility_parents: Optional[Union[List, Tuple, int]] = None,
-        **additional_parameters,
-    ):
-        """Add new input/state node(s) to the neural network.
-
-        Parameters
-        ----------
-        kind :
-            The kind of node to create. If `"continuous-state"` (default), the node will
-            be a regular state node that can have value and/or volatility
-            parents/children. The hierarchical dependencies are specified using the
-            corresponding parameters below. If `"binary-state"`, the node should be the
-            value parent of a binary input. To create an input node, three types of
-            inputs are supported:
-            - `continuous-input`: receive a continuous observation as input.
-            - `binary-input` receives a single boolean as observation. The parameters
-            provided to the binary input node contain: 1. `binary_precision`, the binary
-            input precision, which defaults to `jnp.inf`. 2. `eta0`, the lower bound of
-            the binary process, which defaults to `0.0`. 3. `eta1`, the higher bound of
-            the binary process, which defaults to `1.0`.
-            - `categorical-input` receives a boolean array as observation. The
-            parameters provided to the categorical input node contain: 1.
-            `n_categories`, the number of categories implied by the categorical state.
-
-        .. note::
-           When using a categorical state node, the `binary_parameters` can be used to
-           parametrize the implied collection of binary HGFs.
-
-        .. note:
-           When using `categorical-input`, the implied `n` binary HGFs are automatically
-           created with a shared volatility parent at the third level, resulting in a
-           network with `3n + 2` nodes in total.
-
-        n_nodes :
-            The number of nodes to create (defaults to `1`).
-        node_parameters :
-            Dictionary of parameters. The default values are automatically inferred
-            from the node type. Different values can be provided by passing them in the
-            dictionary, which will overwrite the defaults.
-        value_children :
-            Indexes to the node's value children. The index can be passed as an integer
-            or a list of integers, in case of multiple children. The coupling strength
-            can be controlled by passing a tuple, where the first item is the list of
-            indexes, and the second item is the list of coupling strengths.
-        value_parents :
-            Indexes to the node's value parents. The index can be passed as an integer
-            or a list of integers, in case of multiple children. The coupling strength
-            can be controlled by passing a tuple, where the first item is the list of
-            indexes, and the second item is the list of coupling strengths.
-        volatility_children :
-            Indexes to the node's volatility children. The index can be passed as an
-            integer or a list of integers, in case of multiple children. The coupling
-            strength can be controlled by passing a tuple, where the first item is the
-            list of indexes, and the second item is the list of coupling strengths.
-        volatility_parents :
-            Indexes to the node's volatility parents. The index can be passed as an
-            integer or a list of integers, in case of multiple children. The coupling
-            strength can be controlled by passing a tuple, where the first item is the
-            list of indexes, and the second item is the list of coupling strengths.
-        **kwargs :
-            Additional keyword parameters will be passed and overwrite the node
-            attributes.
-
-        """
-        # extract the node coupling indexes and coupling strengths
-        couplings = []
-        for indexes in [
-            value_children,
-            value_parents,
-            volatility_children,
-            volatility_parents,
-        ]:
-            if indexes is not None:
-                if isinstance(indexes, int):
-                    coupling_idxs = tuple([indexes])
-                    coupling_strengths = tuple([1.0])
-                elif isinstance(indexes, list):
-                    coupling_idxs = tuple(indexes)
-                    coupling_strengths = tuple([1.0] * len(coupling_idxs))
-                elif isinstance(indexes, tuple):
-                    coupling_idxs = tuple(indexes[0])
-                    coupling_strengths = tuple(indexes[1])
-            else:
-                coupling_idxs, coupling_strengths = None, None
-            couplings.append((coupling_idxs, coupling_strengths))
-
-        # create the default parameters set according to the node type
-        if kind == "continuous-state":
-            default_parameters = {
-                "mean": 0.0,
-                "expected_mean": 0.0,
-                "precision": 1.0,
-                "expected_precision": 1.0,
-                "volatility_coupling_children": couplings[2][1],
-                "volatility_coupling_parents": couplings[3][1],
-                "value_coupling_children": couplings[0][1],
-                "value_coupling_parents": couplings[1][1],
-                "tonic_volatility": -4.0,
-                "tonic_drift": 0.0,
-                "autoconnection_strength": 1.0,
-                "observed": 1,
-                "temp": {
-                    "effective_precision": 0.0,
-                    "value_prediction_error": 0.0,
-                    "volatility_prediction_error": 0.0,
-                    "expected_precision_children": 0.0,
-                },
-            }
-        elif kind == "binary-state":
-            default_parameters = {
-                "mean": 0.0,
-                "expected_mean": 0.0,
-                "precision": 1.0,
-                "expected_precision": 1.0,
-                "volatility_coupling_children": couplings[2][1],
-                "volatility_coupling_parents": couplings[3][1],
-                "value_coupling_children": couplings[0][1],
-                "value_coupling_parents": couplings[1][1],
-                "tonic_volatility": 0.0,
-                "tonic_drift": 0.0,
-                "autoconnection_strength": 1.0,
-                "observed": 1,
-                "binary_expected_precision": 0.0,
-                "temp": {
-                    "effective_precision": 0.0,
-                    "value_prediction_error": 0.0,
-                    "volatility_prediction_error": 0.0,
-                    "expected_precision_children": 0.0,
-                },
-            }
-        elif kind == "continuous-input":
-            default_parameters = {
-                "volatility_coupling_parents": None,
-                "value_coupling_parents": None,
-                "input_precision": 1e4,
-                "expected_precision": 1e4,
-                "time_step": 0.0,
-                "value": 0.0,
-                "surprise": 0.0,
-                "observed": 0,
-                "temp": {
-                    "effective_precision": 1.0,  # should be fixed to 1 for input nodes
-                    "value_prediction_error": 0.0,
-                    "volatility_prediction_error": 0.0,
-                },
-            }
-        elif kind == "binary-input":
-            default_parameters = {
-                "expected_precision": jnp.inf,
-                "eta0": 0.0,
-                "eta1": 1.0,
-                "time_step": 0.0,
-                "value": 0.0,
-                "observed": 0,
-                "surprise": 0.0,
-            }
-        elif kind == "categorical-input":
-            if "n_categories" in node_parameters:
-                n_categories = node_parameters["n_categories"]
-            elif "n_categories" in additional_parameters:
-                n_categories = additional_parameters["n_categories"]
-            else:
-                n_categories = 4
-            binary_parameters = {
-                "eta0": 0.0,
-                "eta1": 1.0,
-                "binary_precision": jnp.inf,
-                "n_categories": n_categories,
-                "precision_1": 1.0,
-                "precision_2": 1.0,
-                "precision_3": 1.0,
-                "mean_1": 0.0,
-                "mean_2": -jnp.log(n_categories - 1),
-                "mean_3": 0.0,
-                "tonic_volatility_2": -4.0,
-                "tonic_volatility_3": -4.0,
-            }
-            binary_idxs: List[int] = [
-                1 + i + len(self.edges) for i in range(n_categories)
-            ]
-            default_parameters = {
-                "binary_idxs": binary_idxs,  # type: ignore
-                "n_categories": n_categories,
-                "surprise": 0.0,
-                "kl_divergence": 0.0,
-                "time_step": jnp.nan,
-                "alpha": jnp.ones(n_categories),
-                "pe": jnp.zeros(n_categories),
-                "xi": jnp.array([1.0 / n_categories] * n_categories),
-                "mean": jnp.array([1.0 / n_categories] * n_categories),
-                "value": jnp.zeros(n_categories),
-                "binary_parameters": binary_parameters,
-            }
-
-        if bool(additional_parameters):
-            # ensure that all passed values are valid keys
-            invalid_keys = [
-                key
-                for key in additional_parameters.keys()
-                if key not in default_parameters.keys()
-            ]
-
-            if invalid_keys:
-                raise ValueError(
-                    (
-                        "Some parameter(s) passed as keyword arguments were not found "
-                        f"in the default key list for this node (i.e. {invalid_keys})."
-                        " If you want to create a new key in the node attributes, "
-                        "please use the node_parameters argument instead."
-                    )
-                )
-
-            # if keyword parameters were provided, update the default_parameters dict
-            default_parameters.update(additional_parameters)
-
-        # update the defaults using the dict parameters
-        default_parameters.update(node_parameters)
-        node_parameters = default_parameters
-
-        if "input" in kind:
-            input_type = input_types[kind.split("-")[0]]
-        else:
-            input_type = None
-
-        # convert the structure to a list to modify it
-        edges_as_list: List[AdjacencyLists] = list(self.edges)
-
-        for _ in range(n_nodes):
-            node_idx = len(self.attributes)  # the index of the new node
-
-            # add a new edge
-            edges_as_list.append(
-                AdjacencyLists(
-                    couplings[1][0], couplings[3][0], couplings[0][0], couplings[2][0]
-                )
-            )
-
-            if node_idx == 0:
-                # this is the first node, create the node structure
-                self.attributes = {node_idx: node_parameters}
-                if input_type is not None:
-                    self.inputs = Inputs((node_idx,), (input_type,))
-            else:
-                # update the node structure
-                self.attributes[node_idx] = node_parameters
-
-                if input_type is not None:
-                    # add information about the new input node in the indexes
-                    new_idx = self.inputs.idx
-                    new_idx += (node_idx,)
-                    new_kind = self.inputs.kind
-                    new_kind += (input_type,)
-                    self.inputs = Inputs(new_idx, new_kind)
-
-            # update the existing edge structure so it links to the new node as well
-            for coupling, edge_type in zip(
-                couplings,
-                [
-                    "value_children",
-                    "value_parents",
-                    "volatility_children",
-                    "volatility_parents",
-                ],
-            ):
-                if coupling[0] is not None:
-                    coupling_idxs, coupling_strengths = coupling
-                    for idx, coupling_strength in zip(
-                        coupling_idxs, coupling_strengths  # type: ignore
-                    ):
-                        # unpack this node's edges
-                        (
-                            value_parents,
-                            volatility_parents,
-                            value_children,
-                            volatility_children,
-                        ) = edges_as_list[idx]
-
-                        # update the parents/children's edges depending on the coupling
-                        if edge_type == "value_parents":
-                            if value_children is None:
-                                value_children = (node_idx,)
-                                self.attributes[idx]["value_coupling_children"] = (
-                                    coupling_strength,
-                                )
-                            else:
-                                value_children = value_children + (node_idx,)
-                                self.attributes[idx]["value_coupling_children"] += (
-                                    coupling_strength,
-                                )
-                        elif edge_type == "volatility_parents":
-                            if volatility_children is None:
-                                volatility_children = (node_idx,)
-                                self.attributes[idx]["volatility_coupling_children"] = (
-                                    coupling_strength,
-                                )
-                            else:
-                                volatility_children = volatility_children + (node_idx,)
-                                self.attributes[idx][
-                                    "volatility_coupling_children"
-                                ] += (coupling_strength,)
-                        elif edge_type == "value_children":
-                            if value_parents is None:
-                                value_parents = (node_idx,)
-                                self.attributes[idx]["value_coupling_parents"] = (
-                                    coupling_strength,
-                                )
-                            else:
-                                value_parents = value_parents + (node_idx,)
-                                self.attributes[idx]["value_coupling_parents"] += (
-                                    coupling_strength,
-                                )
-                        elif edge_type == "volatility_children":
-                            if volatility_parents is None:
-                                volatility_parents = (node_idx,)
-                                self.attributes[idx]["volatility_coupling_parents"] = (
-                                    coupling_strength,
-                                )
-                            else:
-                                volatility_parents = volatility_parents + (node_idx,)
-                                self.attributes[idx]["volatility_coupling_parents"] += (
-                                    coupling_strength,
-                                )
-
-                        # save the updated edges back
-                        edges_as_list[idx] = AdjacencyLists(
-                            value_parents,
-                            volatility_parents,
-                            value_children,
-                            volatility_children,
-                        )
-
-        # convert the list back to a tuple
-        self.edges = tuple(edges_as_list)
-
-        # if we are creating a categorical state or state-transition node
-        # we have to generate the implied binary network(s) here
-        if kind == "categorical-input":
-            self = fill_categorical_state_node(
-                self,
-                node_idx=node_idx,
-                binary_input_idxs=node_parameters["binary_idxs"],  # type: ignore
-                binary_parameters=binary_parameters,
-            )
-
-        return self
