+++ src/pyhgf/plots.py
@@ -14,11 +14,11 @@ from pyhgf.typing import input_types
 if TYPE_CHECKING:
     from graphviz.sources import Source
 
-    from pyhgf.model import HGF
+    from pyhgf.model import HGF, Network
 
 
 def plot_trajectories(
-    hgf: "HGF",
+    network: "Network",
     ci: bool = True,
     show_surprise: bool = True,
     show_current_state: bool = False,
@@ -33,8 +33,8 @@ def plot_trajectories(
 
     Parameters
     ----------
-    hgf :
-        An instance of the HGF model.
+    network :
+        An instance of the main Network class.
     ci :
         Show the uncertainty around the values estimates (standard deviation).
     show_surprise :
@@ -121,8 +121,8 @@ def plot_trajectories(
         three_levels_hgf.plot_trajectories();
 
     """
-    trajectories_df = hgf.to_pandas()
-    n_nodes = len(hgf.edges)
+    trajectories_df = network.to_pandas()
+    n_nodes = len(network.edges)
     palette = itertools.cycle(sns.color_palette())
 
     if axs is None:
@@ -130,9 +130,9 @@ def plot_trajectories(
 
     # plot the input node(s)
     # ----------------------
-    for i, input_idx in enumerate(hgf.inputs.idx):
+    for i, input_idx in enumerate(network.inputs.idx):
         plot_nodes(
-            hgf=hgf,
+            network=network,
             node_idxs=input_idx,
             axs=axs[-2 - i],
             show_surprise=show_surprise,
@@ -143,13 +143,13 @@ def plot_trajectories(
 
     # plot continuous and binary nodes
     # --------------------------------
-    ax_i = n_nodes - len(hgf.inputs.idx) - 1
+    ax_i = n_nodes - len(network.inputs.idx) - 1
     for node_idx in range(0, n_nodes):
-        if node_idx not in hgf.inputs.idx:
+        if node_idx not in network.inputs.idx:
             # use different colors for each node
             color = next(palette)
             plot_nodes(
-                hgf=hgf,
+                network=network,
                 node_idxs=node_idx,
                 axs=axs[ax_i],
                 color=color,
@@ -228,13 +228,13 @@ def plot_correlations(hgf: "HGF") -> Axes:
     return ax
 
 
-def plot_network(hgf: "HGF") -> "Source":
+def plot_network(network: "Network") -> "Source":
     """Visualization of node network using GraphViz.
 
     Parameters
     ----------
-    hgf :
-        An instance of the HGF model containing a node structure.
+    network :
+        An instance of main Network class.
 
     Notes
     -----
@@ -257,7 +257,7 @@ def plot_network(hgf: "HGF") -> "Source":
     graphviz_structure.attr("node", shape="circle")
 
     # set input nodes
-    for idx, kind in zip(hgf.inputs.idx, hgf.inputs.kind):
+    for idx, kind in zip(network.inputs.idx, network.inputs.kind):
         if kind == 0:
             label, shape = f"Co-{idx}", "oval"
         elif kind == 1:
@@ -272,13 +272,13 @@ def plot_network(hgf: "HGF") -> "Source":
         )
 
     # create the rest of nodes
-    for i in range(len(hgf.edges)):
+    for i in range(len(network.edges)):
         # only if node is not an input node
-        if i not in hgf.inputs.idx:
+        if i not in network.inputs.idx:
             graphviz_structure.node(f"x_{i}", label=str(i), shape="circle")
 
     # connect value parents
-    for i, index in enumerate(hgf.edges):
+    for i, index in enumerate(network.edges):
         value_parents = index.value_parents
 
         if value_parents is not None:
@@ -289,7 +289,7 @@ def plot_network(hgf: "HGF") -> "Source":
                 )
 
     # connect volatility parents
-    for i, index in enumerate(hgf.edges):
+    for i, index in enumerate(network.edges):
         volatility_parents = index.volatility_parents
 
         if volatility_parents is not None:
@@ -309,7 +309,7 @@ def plot_network(hgf: "HGF") -> "Source":
 
 
 def plot_nodes(
-    hgf: "HGF",
+    network: "Network",
     node_idxs: Union[int, List[int]],
     ci: bool = True,
     show_surprise: bool = True,
@@ -328,8 +328,8 @@ def plot_nodes(
 
     Parameters
     ----------
-    hgf :
-        An instance of the HGF model class after inference over input data.
+    network :
+        An instance of main Network class.
     node_idxs :
         The index(es) of the probabilistic node(s) that should be plotted. If multiple
         indexes are provided, multiple rows will be appended to the figure, one for
@@ -396,7 +396,7 @@ def plot_nodes(
     """
     if not isinstance(node_idxs, list):
         node_idxs = [node_idxs]
-    trajectories_df = hgf.to_pandas()
+    trajectories_df = network.to_pandas()
 
     if axs is None:
         _, axs = plt.subplots(nrows=len(node_idxs), figsize=figsize, sharex=True)
@@ -407,8 +407,8 @@ def plot_nodes(
     for i, node_idx in enumerate(node_idxs):
         # plotting an input node
         # ----------------------
-        if node_idx in hgf.inputs.idx:
-            input_type = hgf.inputs.kind[hgf.inputs.idx.index(node_idx)]
+        if node_idx in network.inputs.idx:
+            input_type = network.inputs.kind[network.inputs.idx.index(node_idx)]
             if input_type == 0:
                 axs[i].scatter(
                     x=trajectories_df.time,
@@ -488,21 +488,21 @@ def plot_nodes(
 
                 # if this is the value parent of an input node
                 # the CI should be treated diffeently
-                if hgf.edges[node_idx].value_children is not None:
+                if network.edges[node_idx].value_children is not None:
                     if np.any(
                         [
                             (
                                 i  # type : ignore
-                                in hgf.edges[  # type: ignore
+                                in network.edges[  # type: ignore
                                     node_idx  # type: ignore
                                 ].value_children  # type: ignore
                             )
                             and kind == 1
-                            for i, kind in enumerate(hgf.inputs.kind)
+                            for i, kind in enumerate(network.inputs.kind)
                         ]
                     ):
                         # get parent node
-                        parent_idx = hgf.edges[  # type: ignore
+                        parent_idx = network.edges[  # type: ignore
                             node_idx  # type: ignore
                         ].value_parents[
                             0
@@ -564,19 +564,19 @@ def plot_nodes(
             # --------------------------------
             if show_observations:
                 # value coupling
-                if hgf.edges[node_idx].value_children is not None:
+                if network.edges[node_idx].value_children is not None:
                     input_colors = plt.cm.cividis(
                         np.linspace(
                             0,
                             1,
-                            len(hgf.edges[node_idx].value_children),  # type: ignore
+                            len(network.edges[node_idx].value_children),  # type: ignore
                         )
                     )
 
                     for ii, child_idx in enumerate(
-                        hgf.edges[node_idx].value_children  # type: ignore
+                        network.edges[node_idx].value_children  # type: ignore
                     ):
-                        if child_idx not in hgf.inputs.idx:
+                        if child_idx not in network.inputs.idx:
                             axs[i].scatter(
                                 trajectories_df.time,
                                 trajectories_df[f"x_{child_idx}_mean"],
@@ -595,9 +595,9 @@ def plot_nodes(
                                 color=input_colors[ii],
                             )
                         else:
-                            child_idx = np.where(np.array(hgf.inputs.idx) == child_idx)[
-                                0
-                            ][0]
+                            child_idx = np.where(
+                                np.array(network.inputs.idx) == child_idx
+                            )[0][0]
                             axs[i].scatter(
                                 trajectories_df.time,
                                 trajectories_df[f"observation_input_{child_idx}"],
@@ -620,7 +620,7 @@ def plot_nodes(
         # plotting surprise
         # -----------------
         if show_surprise:
-            if node_idx in hgf.inputs.idx:
+            if node_idx in network.inputs.idx:
                 node_surprise = trajectories_df[
                     f"observation_input_{node_idx}_surprise"
                 ].to_numpy()
@@ -639,14 +639,16 @@ def plot_nodes(
                     x=trajectories_df.time,
                     y1=node_surprise,
                     y2=node_surprise.min(),
-                    where=hgf.node_trajectories[node_idx]["observed"],
+                    where=network.node_trajectories[node_idx]["observed"],
                     color="#7f7f7f",
                     alpha=0.1,
                     zorder=-1,
                 )
 
                 # hide surprise if the input was not observed
-                node_surprise[hgf.node_trajectories[node_idx]["observed"] == 0] = np.nan
+                node_surprise[network.node_trajectories[node_idx]["observed"] == 0] = (
+                    np.nan
+                )
                 surprise_ax.plot(
                     trajectories_df.time,
                     node_surprise,
